voicemail-service/pyproject.toml
[tool.poetry]
name = "voicemail"
version = "0.1.0"
description = ""
authors = ["Zack Ross <zmr462@usask.ca>"]

[tool.poetry.dependencies]
python = "^3.10"
signalwire = "^2.0.4"
psycopg = {extras = ["binary"], version = "^3.1.3"}
grpcio = "^1.49.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
poetry-dotenv-plugin = "^0.1.0"
grpcio-tools = "^1.49.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
